# api/swagger/swagger.yaml
basePath: /api
info:
  contact:
    email: abdoelrachmad@gmail.com
    name: Maintainer
  description: Simple blog server built with go.
  title: GoBlog Server
  version: "1.0"
paths:
...
  /v1/auth/me:
    get:
      description: Get my user data.
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      createdAt:
                        type: string
                      email:
                        type: string
                      firstName:
                        type: string
                      lastName:
                        type: string
                      roles:
                        items:
                          allOf:
                          - type: object
                          - properties:
                              level:
                                type: integer
                              name:
                                type: string
                              since:
                                type: string
                            type: object
                        type: array
                      uid:
                        type: string
                      updatedAt:
                        type: string
                      username:
                        type: string
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get Me
      tags:
      - Me
    patch:
      consumes:
      - application/json
      - application/msgpack
      description: Update my user data.
      parameters:
      - description: Update me form
        in: body
        name: form
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              email:
                type: string
              firstname:
                type: string
              lastname:
                type: string
              username:
                type: string
            type: object
      produces:
      - application/json
      - application/msgpack
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Me
      tags:
      - Me
    put:
      consumes:
      - application/json
      - application/msgpack
      description: Update my user data.
      parameters:
      - description: Update me form
        in: body
        name: form
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              email:
                type: string
              firstname:
                type: string
              lastname:
                type: string
              username:
                type: string
            type: object
      produces:
      - application/json
      - application/msgpack
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Me
      tags:
      - Me
  /v1/auth/me/password:
    patch:
      consumes:
      - application/json
      - application/msgpack
      description: Update my password.
      parameters:
      - description: Update me password form
        in: body
        name: form
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              newpassword:
                type: string
              newpasswordconfirm:
                type: string
            type: object
      produces:
      - application/json
      - application/msgpack
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Me Password
      tags:
      - Me
    put:
      consumes:
      - application/json
      - application/msgpack
      description: Update my password.
      parameters:
      - description: Update me password form
        in: body
        name: form
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              newpassword:
                type: string
              newpasswordconfirm:
                type: string
            type: object
      produces:
      - application/json
      - application/msgpack
      responses:
        "201":
          description: Created
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update Me Password
      tags:
      - Me
...
securityDefinitions:
  BearerAuth:
    description: '`Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"